+++ src/com/kaylerrenslow/armaDialogCreator/expression/Value.java
@@ -23,6 +23,18 @@ public interface Value {
 			return value;
 		}
 
+		@Override
+		public boolean equals(Object o) {
+			if (o == this) {
+				return true;
+			}
+			if (o instanceof StringLiteral) {
+				StringLiteral other = (StringLiteral) o;
+				return this.value.equals(other.value);
+			}
+			return false;
+		}
+
 		/** @return {@link #getValue()} */
 		@Override
 		public String toString() {
@@ -37,6 +49,19 @@ public interface Value {
 			val = v;
 		}
 
+		@Override
+		public boolean equals(Object o) {
+			if (o == this) {
+				return true;
+			}
+			if (o instanceof NumVal) {
+				NumVal other = (NumVal) o;
+				//see Double.equals()
+				return Double.doubleToLongBits(this.val) == Double.doubleToLongBits(other.val);
+			}
+			return false;
+		}
+
 		public double v() {
 			return val;
 		}
@@ -45,6 +70,18 @@ public interface Value {
 		@Override
 		public String toString() {
 			return ArmaPrecision.format(val);
+		}
+	}
+
+	Void Void = new Void();
+
+	class Void implements Value {
+		private Void() {
+		}
+
+		@Override
+		public String toString() {
+			return "";
 		}
 	}
 }